# ShortUrl.ManagementGui/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "My API",
    "version": "v1"
  },
  "paths": {
    "/Management": {
      "get": {
        "tags": [
          "Management"
        ],
        "operationId": "GetAll",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsAndClaims"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Management"
        ],
        "operationId": "Add",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortUrlModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortUrlModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ShortUrlModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortUrlModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Management/{id}": {
      "get": {
        "tags": [
          "Management"
        ],
        "operationId": "GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortUrlModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Management"
        ],
        "operationId": "DeleteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ShortUrlModel": {
        "required": [
          "key",
          "url"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "key": {
            "maxLength": 10,
            "type": "string"
          },
          "url": {
            "maxLength": 1000,
            "type": "string"
          }
        },
        "nullable": true
      },
      "MyClaim": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "nullable": true
      },
      "ModelsAndClaims": {
        "type": "object",
        "properties": {
          "shortUrlModels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShortUrlModel"
            },
            "nullable": true
          },
          "myClaims": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MyClaim"
            },
            "nullable": true
          }
        },
        "nullable": true
      },
      "Object": {
        "type": "object",
        "nullable": true
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Object"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": {
          "$ref": "#/components/schemas/Object"
        },
        "nullable": true
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token text input below.\r\n\r\nExample: \"Bearer 12345abcdef\"",
        "name": "Autorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}